tidy up with styling and authError pages

+++ Bakery/Controllers/FlavorsController.cs
@@ -1,36 +1,53 @@
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Mvc;
 using Bakery.Models;
-using System.Collections.Generic;
 using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using System.Threading.Tasks;
 using System.Security.Claims;
 
-namespace EpiCodex.Controllers;
+namespace Bakery.Controllers;
 public class FlavorsController: Controller
 {
   private readonly BakeryContext _db;
-  public FlavorsController(BakeryContext db)
+  private readonly UserManager<ApplicationUser> _userManager;
+  public FlavorsController(BakeryContext db, UserManager<ApplicationUser> usrmgr)
   {
     _db = db;
+    _userManager = usrmgr;
   }
   public ActionResult Index()
   {
     return View(_db.Flavors.ToList());
   }
 
-  public ActionResult Details(int id)
-  {
-    Flavor thisFlavor = _db.Flavors.Include(flavor => flavor.JoinEntities).ThenInclude(join => join.Treat)
-                  .FirstOrDefault(flavor => flavor.FlavorId == id);
-    return View(thisFlavor);    
-  }
-  public ActionResult Create()
+  public ActionResult AuthError()
   {
     return View();
+  }
+
+  public ActionResult Details(int id)
+  {
+    Flavor thisFlavor = _db.Flavors
+                        .Include(flavor => flavor.JoinEntities)
+                        .ThenInclude(join => join.Treat)
+                        .FirstOrDefault(flavor => flavor.FlavorId == id);
+    return View(thisFlavor);    
+  }
+
+  public async Task<ActionResult> Create()
+  {
+    string userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+    ApplicationUser currentUser = await _userManager.FindByIdAsync(userId);
+    if(currentUser!=null)
+    {
+      return View();
+    }
+    else 
+    {
+      return RedirectToAction("AuthError");
+    }
   }
   
   [HttpPost]
@@ -41,10 +58,21 @@ public class FlavorsController: Controller
     return RedirectToAction("Index");
   }
 
-  public ActionResult Edit(int id)
+
+  public async Task<ActionResult> Edit(int id)
   {
-    Flavor thisFlavor = _db.Flavors.FirstOrDefault(flav => flav.FlavorId == id);
-    return View(thisFlavor);
+    string userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+    ApplicationUser currentUser = await _userManager.FindByIdAsync(userId);
+    if(currentUser!=null)
+    {
+      Flavor thisFlavor = _db.Flavors.FirstOrDefault(flav => flav.FlavorId == id);
+      return View(thisFlavor);
+    }
+    else 
+    {
+      return RedirectToAction("AuthError");
+    }
+    
   }
 
   [HttpPost]
@@ -55,10 +83,21 @@ public class FlavorsController: Controller
     return RedirectToAction("Index");
   }
 
-  public ActionResult Delete(int id)
+
+  public async Task<ActionResult> Delete(int id)
   {
-    Flavor flav = _db.Flavors.FirstOrDefault(flav => flav.FlavorId == id);
-    return View(flav);
+    string userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+    ApplicationUser currentUser = await _userManager.FindByIdAsync(userId);
+    if(currentUser!=null)
+    {
+      Flavor flav = _db.Flavors.FirstOrDefault(flav => flav.FlavorId == id);
+      return View(flav);
+    }
+    else 
+    {
+      return RedirectToAction("AuthError");
+    }
+    
   }
 
   [HttpPost, ActionName("Delete")]
@@ -71,12 +110,22 @@ public class FlavorsController: Controller
   }
 
   [HttpPost]
-  public ActionResult DeleteJoin(int joinId)
+  public async Task<ActionResult> DeleteJoin(int joinId)
   {
-    FlavorTreat joinEntry = _db.FlavorTreats.FirstOrDefault(entry => entry.FlavorTreatId == joinId);
-    _db.FlavorTreats.Remove(joinEntry);
-    _db.SaveChanges();
-    return RedirectToAction("Index");
+    string userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+    ApplicationUser currentUser = await _userManager.FindByIdAsync(userId);
+    if(currentUser!=null)
+    {
+      FlavorTreat joinEntry = _db.FlavorTreats.FirstOrDefault(entry => entry.FlavorTreatId == joinId);
+      _db.FlavorTreats.Remove(joinEntry);
+      _db.SaveChanges();
+      return RedirectToAction("Index");
+    }
+    else 
+    {
+      return RedirectToAction("AuthError");
+    }
+    
   }
 
 }
